Refactor - Prep CSS for expanding modifications
What
=================
Prep for expanding beyond modifications of the blog to the conference
page, possibly more.

Why
=================
It would be misleading to leave the filename referencing "blog" when it
modifies other elements of the site as well. Mod was the term used as it
is generic enough to describe the fact that these changes happened after
the initial release of the theme to my client.

Ticket: N/A